Add enforced culling if lice count is excessive
