+++ slim/simulation/organisation.py
@@ -14,7 +14,7 @@ import ray
 from numpy.random import SeedSequence
 from ray.util.queue import Queue as RayQueue
 
-from slim.types.policies import no_observation
+from slim.types.policies import no_observation, agent_to_id, FALLOW
 from slim.types.queue import *
 from .farm import (
     FarmActor,
@@ -37,6 +37,9 @@ if TYPE_CHECKING:
 
 
 class FarmPool(ABC):
+    def __init__(self):
+        self.reported_aggregations = {}
+
     @abstractmethod
     def start(self):
         pass
@@ -55,18 +58,48 @@ class FarmPool(ABC):
     ):
         pass
 
-    def handle_aggregation_rate(self, spaces: SimulatorSpace, threshold: float) -> bool:
-        to_broadcast = False
-        for space in spaces.values():
-            if space["reported_aggregation"].item() >= threshold:
-                to_broadcast = True
+    def handle_aggregation_rate(
+        self, spaces: SimulatorSpace, threshold: float, limit: float, weeks: int
+    ) -> Tuple[bool, List[str]]:
+        """
+        Handle the aggregation rate of the specific cages. This has two functions:
+        - suggest all farmers to apply a treatment
+        - force a fallowing to non-compliant farms
 
-        return to_broadcast
+        This implements a simplified approach to the Scottish's CoGP Plan.
+        The algorithm is the following:
+
+        - if (lice count of a farm f >= org_aggregation_threshold):
+            recommend treatment to everyone
+        - if (lice count of a farm f) >= cogp_aggregation_threshold
+            count as a strike
+        - if S strikes have been reached for a farm f:
+            enforce culling on farm f
+
+        :returns:
+        """
+        to_broadcast = False
+        to_cull = []
+        for farm, space in spaces.items():
+            aggregation = space["reported_aggregation"].item()
+            if aggregation >= threshold:
+                to_broadcast = True
+            past_counts = self.reported_aggregations.setdefault(farm, [])
+            if aggregation > 0:
+                past_counts.append(aggregation)
+            if len(past_counts) >= weeks and all(
+                count >= limit for count in past_counts[-weeks:]
+            ):
+                to_cull.append(farm)
+                past_counts.clear()
+
+        return to_broadcast, to_cull
 
 
 class SingleProcFarmPool(FarmPool):
     # A single-process farm pool. Useful for debugging
     def __init__(self, cfg: Config, *args):
+        super().__init__()
         nfarms = cfg.nfarms
         self.threshold = cfg.aggregation_rate_threshold
         self.cfg = cfg
@@ -139,6 +172,7 @@ class SingleProcFarmPool(FarmPool):
 class MultiProcFarmPool(FarmPool):
     # A multi-process farm pool.
     def __init__(self, cfg: Config, *args, **kwargs):
+        super().__init__()
         self.cfg = cfg
         self._extra_farm_args = args
 
@@ -365,6 +399,11 @@ class Organisation:
         :returns: the cumulated reward from all the farm updates, and logs generated by all the farms
         """
 
+        for farm in self.to_cull:
+            actions[agent_to_id(farm)] = FALLOW
+
+        self.to_cull = []
+
         offspring_per_farm, payoffs, logs = self.farms.step(
             cur_date, actions, *self.get_external_pressure()
         )
@@ -373,9 +412,13 @@ class Organisation:
         self.update_offspring_average(offspring_per_farm)
         self.update_genetic_ratios(self.averaged_offspring)
 
-        to_treat = self.farms.handle_aggregation_rate(
-            spaces, self.cfg.aggregation_rate_threshold
+        to_treat, self.to_cull = self.farms.handle_aggregation_rate(
+            spaces,
+            self.cfg.aggregation_rate_threshold,
+            self.cfg.aggregation_rate_enforcement_limit,
+            self.cfg.aggregation_rate_enforcement_weeks,
         )
+
         for space in spaces.values():
             space["asked_to_treat"] = np.array([int(to_treat)], dtype=np.int8)
 
@@ -392,6 +435,7 @@ class Organisation:
             self.cfg.initial_genetic_ratios
         )
         self.averaged_offspring = empty_geno_from_cfg(self.cfg)
+        self.to_cull = []
 
     def stop(self):
         self.farms.stop()
